feature: clean up and add client functions for tasks
